deprecated and disabled old high opmodes, made some reminders in right auto and removed P/D telemetry for lift
