+++ TeamCode/src/main/kotlin/org/firstinspires/ftc/teamcodekt/opmodes/teleop/LiftTestingOp.kt
@@ -1,11 +1,13 @@
 package org.firstinspires.ftc.teamcodekt.opmodes.teleop
 
 import com.acmerobotics.dashboard.config.Config
+import com.qualcomm.robotcore.eventloop.opmode.Disabled
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp
 import org.firstinspires.ftc.teamcode.AutoData.*
 
 
 @Config
+@Disabled
 @TeleOp
 class LiftTestingOp : RogueBaseTele() {
